Extract static data from Region into StaticRegion
