+++ agot-bg-game-server/src/client/MapComponent.tsx
@@ -2,7 +2,6 @@ import {Component, ReactNode} from "react";
 import GameClient from "./GameClient";
 import IngameGameState from "../common/ingame-game-state/IngameGameState";
 import * as React from "react";
-import Border from "../common/ingame-game-state/game-data-structure/Border";
 import Region from "../common/ingame-game-state/game-data-structure/Region";
 import Unit from "../common/ingame-game-state/game-data-structure/Unit";
 import PlanningGameState from "../common/ingame-game-state/planning-game-state/PlanningGameState";
@@ -23,6 +22,7 @@ import BetterMap from "../utils/BetterMap";
 import _ from "lodash";
 import PartialRecursive from "../utils/PartialRecursive";
 import joinReactNodes from "./utils/joinReactNodes";
+import StaticBorder from "../common/ingame-game-state/game-data-structure/static-data-structure/StaticBorder";
 
 interface MapComponentProps {
     gameClient: GameClient;
@@ -261,7 +261,7 @@ export default class MapComponent extends Component<MapComponentProps> {
         );
     }
 
-    getBorderPathD(border: Border): string {
+    getBorderPathD(border: StaticBorder): string {
         return border.polygon.map(p => p.x + "," + p.y).join(" ");
     }
