+++ agot-bg-game-server/src/common/ingame-game-state/game-data-structure/static-data-structure/StaticWorld.ts
@@ -1,0 +1,13 @@
+import StaticRegion from "./StaticRegion";
+import BetterMap from "../../../../utils/BetterMap";
+import StaticBorder from "./StaticBorder";
+
+export default class StaticWorld {
+    staticRegions: BetterMap<string, StaticRegion>;
+    staticBorders: StaticBorder[];
+
+    constructor(staticRegions: BetterMap<string, StaticRegion>, staticBorders: StaticBorder[]) {
+        this.staticRegions = staticRegions;
+        this.staticBorders = staticBorders;
+    }
+}
